+++ src/lsp/completions/tests/mod.rs
@@ -8,6 +8,7 @@ mod account_alias_tests;
 mod account_loader_tests;
 mod handler_call_return_tests;
 mod handler_expression_tests;
+mod handler_iterator_tests;
 mod handler_member_tests;
 mod handler_method_return_tests;
 mod handler_method_tests;